requirements.txt
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
Authlib==1.6.6
backports.tarfile==1.2.0
beartype==0.22.9
cachetools==7.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
cryptography==46.0.4
cyclopts==4.5.1
diskcache==5.6.3
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
fakeredis==2.33.0
fastmcp==2.14.4
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
lupa==2.6
markdown-it-py==4.0.0
mcp==1.26.0
mdurl==0.1.2
more-itertools==10.8.0
oauthlib==3.3.1
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.0
pathable==0.4.4
pathvalidate==3.3.1
platformdirs==4.5.1
prometheus_client==0.24.1
proto-plus==1.27.0
protobuf==6.33.5
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydocket==0.16.6
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
pyperclip==1.11.0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
redis==7.1.0
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
rich-rst==1.3.2
rpds-py==0.30.0
rsa==4.9.1
shellingham==1.5.4
sortedcontainers==2.4.0
sse-starlette==3.2.0
starlette==0.52.1
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
websockets==16.0
wrapt==1.17.3
zipp==3.23.0
